Backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
langchain==0.1.0
langchain-openai==0.0.2
langchain-neo4j==0.0.3
langchain-community==0.0.10
langchain-core==0.1.0
openai==1.6.1
neo4j==5.15.0
pymupdf==1.23.14
pydantic==2.5.2